requirements.txt
Django==1.10
django-picklefield==0.3.2
django-downloadview==1.9
django-webpack-loader==0.3.3
djangorestframework==3.4.3
djangorestframework-jwt==1.8.0
drf-nested-routers==0.11.1
gunicorn==19.6.0
Pillow==3.4.1
psycopg2==2.6.2
PyJWT==1.4.2
requests==2.11.1
six==1.10.0
stripe==1.41.1
paypalrestsdk==1.12.0
django-rest-auth==0.8.2
django-allauth==0.29.0
disqus-python==0.4.2
